Refactored skill sets sql.
